refactor: 프로필 대시보드 컴포넌트 구조 분리
